@@ -5,7 +5,7 @@ author: Bruno Jurković
 homepage: https://github.com/BrunoJurkovic/popup_card
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
