@@ -6,10 +6,10 @@ import 'package:popup_card/src/hero_route.dart';
 
 // [PopupItemLauncher], when wrapped around a widget like an [Icon], launches the [PopUpItem] widget.
 class PopupItemLauncher extends StatelessWidget {
-  final Object tag;
-  final Widget child;
-  final Widget popUp;
-  const PopupItemLauncher({Key key, this.tag, this.child, this.popUp})
+  final Object? tag;
+  final Widget? child;
+  final Widget? popUp;
+  const PopupItemLauncher({Key? key, this.tag, this.child, this.popUp})
       : super(key: key);
 
   @override
@@ -19,12 +19,12 @@ class PopupItemLauncher extends StatelessWidget {
       child: GestureDetector(
         onTap: () {
           Navigator.of(context).push(HeroDialogRoute(builder: (context) {
-            return popUp;
+            return popUp!;
           }));
         },
         child: Hero(
-          tag: tag,
-          child: child,
+          tag: tag!,
+          child: child!,
         ),
       ),
     );
@@ -34,20 +34,20 @@ class PopupItemLauncher extends StatelessWidget {
 // This is the actual pop up card. You provide this to [PopUpItemLauncher].
 class PopUpItem extends StatelessWidget {
   final Object tag;
-  final Widget builder;
+  final Widget child;
   final Color color;
   final double elevation;
   final EdgeInsetsGeometry padding;
   final ShapeBorder shape;
-  const PopUpItem(
-      {Key key,
-      this.tag,
-      this.builder,
-      this.color,
-      this.elevation,
-      this.padding,
-      this.shape})
-      : super(key: key);
+  PopUpItem({
+    Key? key,
+    required this.tag,
+    required this.child,
+    required this.color,
+    required this.padding,
+    required this.shape,
+    this.elevation = 2,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -63,7 +63,7 @@ class PopUpItem extends StatelessWidget {
             child: SingleChildScrollView(
               child: Padding(
                 padding: padding,
-                child: builder,
+                child: child,
               ),
             ),
           ),
